# swagger/swagger.yaml
basePath: /
definitions:
...
  localai.APIInstructionResponse:
    properties:
      description:
        type: string
      name:
        type: string
      swagger_fragment:
        additionalProperties: {}
        type: object
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    name: LocalAI
    url: https://localai.io
  description: The LocalAI Rest API.
  license:
    name: MIT
    url: https://raw.githubusercontent.com/mudler/LocalAI/master/LICENSE
  title: LocalAI API
  version: 2.0.0
paths:
...
  /api/instructions/{name}:
    get:
      description: Returns a markdown guide (default) or filtered OpenAPI fragment
        (format=json) for a named instruction
      parameters:
      - description: Instruction name (e.g. chat-inference, config-management)
        in: path
        name: name
        required: true
        type: string
      - description: 'Response format: json for OpenAPI fragment, omit for markdown'
        in: query
        name: format
        type: string
      produces:
      - application/json
      - text/markdown
      responses:
        "200":
          description: instruction documentation
          schema:
            $ref: '#/definitions/localai.APIInstructionResponse'
        "404":
          description: instruction not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an instruction's API guide or OpenAPI fragment
      tags:
      - instructions
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"